Consolidate d.ts files to a single file per DefinitelyTyped guidelines
