--- src/aio.d.ts
@@ -1,7 +1,0 @@
-export class Aio {
-    public constructor(pin: number);
-    public read(): number;
-    public readFloat(): number;
-    public setBit(bits: number): number;
-    public getBit(): number;
-}
