--- src/common.d.ts
@@ -1,14 +1,0 @@
-export const SUCCESS: number;
-export const ERROR_FEATURE_NOT_IMPLEMENTED: number;
-export const ERROR_FEATURE_NOT_SUPPORTED: number;
-export const ERROR_INVALID_VERBOSITY_LEVEL: number;
-export const ERROR_INVALID_PARAMETER: number;
-export const ERROR_INVALID_HANDLE: number;
-export const ERROR_NO_RESOURCES: number;
-export const ERROR_INVALID_RESOURCE: number;
-export const ERROR_INVALID_QUEUE_TYPE: number;
-export const ERROR_NO_DATA_AVAILABLE: number;
-export const ERROR_INVALID_PLATFORM: number;
-export const ERROR_PLATFORM_NOT_INITIALISED: number;
-export const ERROR_PLATFORM_ALREADY_INITIALISED: number;
-export const ERROR_UNSPECIFIED: number;
