--- src/iio.d.ts
@@ -1,5 +1,0 @@
-export class Iio {
-    public constructor(device: number);
-    public getDeviceName(): string;
-    public registerEventHandler(handler: Iio): void;
-}
